+++ src/main/java/com/jasik/momsnaggingapi/domain/release/controller/ReleaseController.java
@@ -1,0 +1,7 @@
+package com.jasik.momsnaggingapi.domain.release.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ReleaseController {
+}
